Consolidate python binding
